feat: add admin system

+++ app/api/property/route.ts
@@ -1,10 +1,14 @@
 import { NextRequest, NextResponse } from "next/server";
-import { UTApi } from "uploadthing/server";
+import { utapi } from "@/lib/uploadthing-server";
+import { auth } from "@clerk/nextjs";
 
 import { db } from "@/lib/prisma";
 import { propertySchema } from "@/lib/validations";
+import { getIsAdmin } from "@/app/actions";
 
 export async function POST(req: NextRequest) {
+  if (!(await isAdmin()))
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
   const body = await req.json();
   const parsedBody = propertySchema.safeParse(body);
 
@@ -19,15 +23,22 @@ export async function POST(req: NextRequest) {
 }
 
 export async function DELETE(req: NextRequest) {
+  if (!(await isAdmin()))
+    return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
   const { id } = await req.json();
   const property = await db.property.findUnique({
     select: { imageKeys: true },
     where: { id },
   });
   if (property && property.imageKeys.length) {
-    const utapi = new UTApi();
     await utapi.deleteFiles(property.imageKeys);
   }
   await db.property.delete({ where: { id } });
   return NextResponse.json({ message: "Property deleted" });
+}
+
+async function isAdmin() {
+  const { userId } = auth();
+  if (!(await getIsAdmin(userId))) return false;
+  return true;
 }
